+++ lib/csr_matrix.cpp
@@ -1,11 +1,11 @@
 #include "csr_matrix.hpp"
 
-csr_matrix::csr_matrix(int size, std::set<std::pair<int, int>> coords) : size(size), col_idx(), row_ptr() {
+csr_matrix::csr_matrix(size_t size, std::set<std::pair<size_t, size_t>> coords) : size(size), col_idx(), row_ptr() {
     col_idx.reserve(coords.size());
     row_ptr.reserve(size + 1);
-    for (int y = 0; y < size; y++) {
+    for (size_t y = 0; y < size; y++) {
         bool first_nz = true;
-        for (int x = 0; x < size; x++) { // TODO: c++20 -> set.contains
+        for (size_t x = 0; x < size; x++) { // TODO: c++20 -> set.contains
             if (coords.count(std::make_pair(x, y))) {
                 if (first_nz) { // TODO: unwrap this into two partial loops instead
                     row_ptr.push_back(col_idx.size());
@@ -15,7 +15,7 @@ csr_matrix::csr_matrix(int size, std::set<std::pair<int, int>> coords) : size(si
             }
         }
         if (first_nz) {
-            int back = row_ptr.back();
+            size_t back = row_ptr.back();
             row_ptr.push_back(back);
         }
     }
